Revert "ushort change"
This reverts commit 4a898a16505b103a1d83ea036975f453f947580f.

+++ neo-cli/Shell/MainService.cs
@@ -110,8 +110,8 @@ namespace Neo.Shell
                 return true;
             }
 
-            ushort count;
-            if (args.Length < 3 || !ushort.TryParse(args[2], out count) || count == 0)
+            byte count;
+            if (args.Length < 3 || !byte.TryParse(args[2], out count) || count == 0)
                 count = 1;
 
             List<string> addresses = new List<string>();
